Wiktor test początkie

+++ src/main/java/com/example/przewozy/repo/KlientRepository.java
@@ -2,7 +2,7 @@ package com.example.przewozy.repo;
 import com.example.przewozy.entity.Klient;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface KlientRepository extends JpaRepository<Klient, Long> {
+public interface KlientRepository extends JpaRepository<Klient, Integer> {
     List<Klient> findByNazwisko(String nazwisko);
 }
